In appspec.yml stop_application script modified.

+++ scripts/stop_application.sh
@@ -1,2 +1,10 @@
 #!/bin/bash
-pkill node
+if pgrep node > /dev/null
+then
+    echo "Stopping Node.js application..."
+    pkill node
+    exit 0
+else
+    echo "No Node.js application is currently running."
+    exit 0
+fi
